Replaces references to Shadowsocks
