+++ client/go/outline/client.go
@@ -15,7 +15,14 @@
 package outline
 
 import (
+	"fmt"
+	"net"
+
+	"github.com/Jigsaw-Code/outline-apps/client/go/outline/connectivity"
+	"github.com/Jigsaw-Code/outline-apps/client/go/outline/platerrors"
 	"github.com/Jigsaw-Code/outline-sdk/transport"
+	"github.com/Jigsaw-Code/outline-sdk/transport/shadowsocks"
+	"github.com/eycorsican/go-tun2socks/common/log"
 )
 
 // Client provides a transparent container for [transport.StreamDialer] and [transport.PacketListener]
@@ -24,4 +31,86 @@ import (
 type Client struct {
 	transport.StreamDialer
 	transport.PacketListener
+}
+
+// NewClient creates a new Outline client from a configuration string.
+func NewClient(transportConfig string) (*Client, error) {
+	config, err := parseConfigFromJSON(transportConfig)
+	if err != nil {
+		return nil, newIllegalConfigErrorWithDetails("Shadowsocks config must be a valid JSON string", ".", transportConfig, "JSON string", err)
+	}
+	prefixBytes, err := ParseConfigPrefixFromString(config.Prefix)
+	if err != nil {
+		return nil, err
+	}
+	return newShadowsocksClient(config.Host, int(config.Port), config.Method, config.Password, prefixBytes)
+}
+
+func newShadowsocksClient(host string, port int, cipherName, password string, prefix []byte) (*Client, error) {
+	if err := validateConfig(host, port, cipherName, password); err != nil {
+		return nil, err
+	}
+
+	// TODO: consider using net.LookupIP to get a list of IPs, and add logic for optimal selection.
+	proxyAddress := net.JoinHostPort(host, fmt.Sprint(port))
+
+	cryptoKey, err := shadowsocks.NewEncryptionKey(cipherName, password)
+	if err != nil {
+		return nil, newIllegalConfigErrorWithDetails("cipher&password pair is not valid",
+			"cipher|password", cipherName+"|"+password, "valid combination", err)
+	}
+
+	// We disable Keep-Alive as per https://datatracker.ietf.org/doc/html/rfc1122#page-101, which states that it should only be
+	// enabled in server applications. This prevents the device from unnecessarily waking up to send keep alives.
+	streamDialer, err := shadowsocks.NewStreamDialer(&transport.TCPEndpoint{Address: proxyAddress, Dialer: net.Dialer{KeepAlive: -1}}, cryptoKey)
+	if err != nil {
+		return nil, platerrors.PlatformError{
+			Code:    platerrors.SetupTrafficHandlerFailed,
+			Message: "failed to create TCP traffic handler",
+			Details: platerrors.ErrorDetails{"proxy-protocol": "shadowsocks", "handler": "tcp"},
+			Cause:   platerrors.ToPlatformError(err),
+		}
+	}
+	if len(prefix) > 0 {
+		log.Debugf("Using salt prefix: %s", string(prefix))
+		streamDialer.SaltGenerator = shadowsocks.NewPrefixSaltGenerator(prefix)
+	}
+
+	packetListener, err := shadowsocks.NewPacketListener(&transport.UDPEndpoint{Address: proxyAddress}, cryptoKey)
+	if err != nil {
+		return nil, platerrors.PlatformError{
+			Code:    platerrors.SetupTrafficHandlerFailed,
+			Message: "failed to create UDP traffic handler",
+			Details: platerrors.ErrorDetails{"proxy-protocol": "shadowsocks", "handler": "udp"},
+			Cause:   platerrors.ToPlatformError(err),
+		}
+	}
+
+	return &Client{StreamDialer: streamDialer, PacketListener: packetListener}, nil
+}
+
+// Error number constants exported through gomobile
+const (
+	NoError                     = 0
+	Unexpected                  = 1
+	NoVPNPermissions            = 2 // Unused
+	AuthenticationFailure       = 3
+	UDPConnectivity             = 4
+	Unreachable                 = 5
+	VpnStartFailure             = 6  // Unused
+	IllegalConfiguration        = 7  // Electron only
+	ShadowsocksStartFailure     = 8  // Unused
+	ConfigureSystemProxyFailure = 9  // Unused
+	NoAdminPermissions          = 10 // Unused
+	UnsupportedRoutingTable     = 11 // Unused
+	SystemMisconfigured         = 12 // Electron only
+)
+
+// CheckConnectivity determines whether the Shadowsocks proxy can relay TCP and UDP traffic under
+// the current network. Parallelizes the execution of TCP and UDP checks, selects the appropriate
+// error code to return accounting for transient network failures.
+// Returns an error if an unexpected error ocurrs.
+func CheckConnectivity(client *Client) (int, error) {
+	errCode, err := connectivity.CheckConnectivity(client, client)
+	return errCode.Number(), err
 }
