+++ client/go/outline/client_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package shadowsocks
+package outline
 
 import "testing"
 
@@ -68,7 +68,7 @@ func Test_NewClientFromJSON_Errors(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			got, err := NewClientFromJSON(tt.input)
+			got, err := NewClient(tt.input)
 			if err == nil || got != nil {
 				t.Errorf("NewClientFromJSON() expects an error, got = %v", got)
 				return
