+++ client/src/cordova/android/OutlineAndroidLib/outline/src/main/aidl/org/outline/shadowsocks/ShadowsocksConfig.aidl
@@ -12,9 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package org.outline.shadowsocks;
+package org.outline;
 
-parcelable ShadowsocksConfig {
+parcelable OutlineConfig {
   String host;
   int port;
   String password;
