+++ client/src/cordova/apple/OutlineLib/VpnExtension/Sources/PacketTunnelProvider.h
@@ -34,7 +34,7 @@ typedef NS_ENUM(NSInteger, ErrorCode) {
   serverUnreachable = 5,
   vpnStartFailure = 6,
   illegalServerConfiguration = 7,
-  shadowsocksStartFailure = 8,
+  clientStartFailure = 8,
   configureSystemProxyFailure = 9,
   noAdminPermissions = 10,
   unsupportedRoutingTable = 11,
